requirements_api.txt
flask>=3.0.0
gunicorn>=21.0.0
azure-storage-blob>=12.19.0
torch>=2.0.0
scikit-learn>=1.2.0
numpy>=1.24.0
python-dotenv>=1.0.0